refactor: split typeutils and asset identities from the data stack
Closes #127
