+++ hyperion/infrastructure/geo/gmaps.py
@@ -8,7 +8,7 @@ from typing import Any, ClassVar
 import googlemaps
 
 from hyperion.config import geo_config
-from hyperion.entities.catalog import PersistentStoreAsset
+from hyperion.domain.assets import PersistentStoreAsset
 from hyperion.infrastructure.cache import PersistentCache
 from hyperion.infrastructure.geo.location import Location, NamedLocation
 from hyperion.log import get_logger
